Improve noise reduction
Moved noise reduction to it's own function that receives an image and
returns a (hopefully) noiseless image.

It checks every component looking if it's in range of reasonable height,
width and pixel count. This values are hardcode for now (will probably
remain hardcode, but in another file).

Also, using os.path to make image save path work for everybody.