Create options file
options file stores information about the screen portion to grab and the
ranges for a reseanable pixel component for noise reduction.

The commit file is just a sample and you should make a copy of it and
rename it `options.py`, as stated on the file itself.

Just so there are no accidental commits of the actual options file, it
was added to .gitignore