Change validation logic
Now is_valid_text check not only if there are three lines on the output,
but if each of those lines forms a valid number.

Main loop is also changed so not to get stuck in some bases with noise
around the loot image. get_loot function checks for a valid loot, if it
finds a valid loot it returns, if the loot is invalid because the parsed
text is empty, it tries again indefinetly. If the loot is invalid
because the format is wrong it tries again up to three times and gives
up, skipping to the next base.