Use tesserocr instead of pytesseract
tesserocr allows getting the confidence level of each individual symbol
from the image. Using this instead of having to mess with precision on
the noise reduction values.

Now noise reduction only checks for the height of a component, which is
easier because there are 3 rects now, one for each loot type.

tesserocr gets rid of anything that has confidence level below 75% and
just concats everything with confidence above that.

Using PSM.SINGLE_LINE now that each loot type has it's own rect.