Add unloaded eagle recognition
Using cv2 to find an unloaded eagle artillery on the base and provide it
as a `unloaded` variable to be considered when deciding if the program
should halt.

The decision process to find if an eagle is unloaded is to first find a
good enough match (> 0.6) of an unloaded eagle and then check if this
match is better than the best possible match of a loaded eagle.

The scale config option can be a bit fiddly, there might be a need to
add a program to generate this scale automatically.